nodes-lib: separate export entrypoints for browser/node use to prevent native node binary deps
