+++ nodes-lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@desci-labs/nodes-lib",
-  "version": "0.2.0-rc2",
+  "version": "0.2.0-rc3",
   "homepage": "https://github.com/desci-labs/nodes#readme",
   "description": "Stand-alone client library for interacting with desci-server",
   "repository": {
@@ -15,6 +15,16 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "type": "module",
+  "exports": {
+    "./browser": {
+      "import": "./dist/export-entrypoints/browser.js",
+      "types": "./dist/export-entrypoints/browser.d.ts"
+    },
+    "./node": {
+      "import": "./dist/export-entrypoints/node.js",
+      "types": "./dist/export-entrypoints/node.d.ts"
+    }
+  },
   "files": [
     "dist"
   ],
