+++ nodes-lib/src/export-entrypoints/browser.ts
@@ -1,0 +1,28 @@
+/**
+ * Browser-safe exports from nodes-lib
+ * This module excludes functions that depend on Node.js-specific modules or FlightSQL
+ */
+
+// Re-export everything from api.ts
+// (uploadFiles and getPublishHistory have been moved to node-only modules)
+export * from "../shared/api.js";
+
+export * from "../shared/publish.js";
+
+export * from "../shared/config/index.js";
+
+export * from "../shared/errors.js";
+
+export * from "../shared/util/headers.js";
+export * from "../shared/util/signing.js";
+export * from "../shared/util/converting.js";
+export * from "../shared/util/manifest.js";
+
+export * from "../shared/chain.js";
+
+// Re-export Codex operations that don't use FlightSQL
+// (getFullState, getCurrentState, getCodexHistory have been moved to node-only modules)
+export {
+  codexPublish,
+  getStreamController
+} from "../shared/codex.js";
