+++ nodes-lib/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./api.js"
-export * from "./config/index.js";
-export * from "./errors.js";
-export { getHeaders } from "./util/headers.js";
+export * from "./shared/api.js"
+export * from "./shared/config/index.js";
+export * from "./shared/errors.js";
+export { getHeaders } from "./shared/util/headers.js";
