+++ nodes-lib/src/shared/abi/DpidRegistry.json
