+++ nodes-lib/src/shared/abi/ResearchObject.json
