+++ nodes-lib/src/shared/automerge.ts
