+++ nodes-lib/src/shared/chain.ts
