+++ nodes-lib/src/shared/config/chain.ts
