+++ nodes-lib/src/shared/errors.ts
