+++ nodes-lib/src/shared/publish.ts
