+++ nodes-lib/src/shared/util/converting.ts
