+++ nodes-lib/src/shared/util/signing.ts
